% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{argyle}
\alias{argyle}
\alias{argyle-package}
\title{argyle: An \code{R} package for import and QC of genotypes from Illumina Infinium arrays}
\description{
argyle: An \code{R} package for import and QC of genotypes from Illumina Infinium arrays
}
\section{The \code{genotypes} object}{

The \code{genotypes} class is just a matrix (sites x samples) with row and column names, and a dataframe
(in \code{attr(,"map")}) describing marker positions.  Nearly all the functions in this package expect
a \code{genotypes} object as input.  For details, see \code{\link{genotypes}}.
}

\section{Accessing metadata}{

The \code{`$`} operator is overloaded for the \code{genotypes} class, so that writing \code{attr(g, "normalized")}
is equivalent to writing \code{g$normalized}.  But this works only in one direction: \code{g$normalized <- FALSE}
fails.  Nefarious users can modify attributes directly using the standard and somewhat convoluted syntax
\code{attr(,"x") <- y} but do so at their own risk.  For safety, always check that the resulting object remains
valid (all internal parts having matching dimensions and names) with a call to \code{validate(g)}.

Accessor functions are provided for the marker map (\code{markers(g)}), sample metadata (\code{samples(g)}),
and intensity matrices (\code{intensity(g)}).
}

\section{Notes on allele encoding}{

For the purposes of this package, all markers on an array are treated as biallelic SNPs, and all samples are assumed
to be diploid for the autosomes. Genotype calls are reported by Illumina BeadStudio as a two-character vector of
nucleotides: eg. \code{AA}, \code{AG}, \code{GG} for an [A/G] SNP.  The \code{-} character indicates a missing call
("no-call").  On import, these calls are summarized to a single character, one of \code{ACGTHN}
(\code{H} = heterozygous, \code{N} = no-call).

For most analyses a numeric representation of genotypes is desirable.  The function \code{recode.genotypes()}
performs this conversion.  When reference alleles are provided in columns "A1" (REF) and "A2" (ALT) in the marker map,
genotypes are recoded 0 (homozygous REF), 1 (heterozygous), 2 (homozygous ALT) or \code{NA} (missing).

Recoding can also be "relative": that is, performed with respect to the major and minor allele as defined by
the dataset itself.  In this case the recoded 0 (homozygous major allele), 1 (heterozygous), 2 (homozygous
minor allele) or \code{NA} (missing).

The attribute \code{alleles} tracks the current allele encoding.
}

