% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{as.data.frame.genotypes}
\alias{as.data.frame.genotypes}
\title{Convert genotypes to a dataframe}
\usage{
\method{as.data.frame}{genotypes}(gty, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{...}{ignored}
}
\value{
a \code{data.frame} with marker information in the leftmost columns, followed by a
	matrix of genotypes with samples in columns
}
\description{
Convert genotypes to a dataframe
}
\details{
In general the dataframe will be a less-efficient way to store genotypes, but is a
	useful intermediate for writing genotypes to disk in a human-readable format.
}

