% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{as.genotypes.cross}
\alias{as.genotypes.cross}
\title{Convert an \code{R/qtl} object to a \code{genotypes} object}
\usage{
\method{as.genotypes}{cross}(x, ...)
}
\arguments{
\item{x}{a \code{qtl::cross} object}

\item{...}{ignored}
}
\value{
an object of class \code{genotypes}
}
\description{
Convert an \code{R/qtl} object to a \code{genotypes} object
}
\details{
Karl Broman's \code{R/qtl} is a widely-used package for mapping quantiative traits
	in experimental crosses of laboratory organisms and crop plants.  It expects genotypes to
	be coded with respect to parental lines: eg. AA, AB, BB for an F2 cross between (true-breeding)
	lines A and B.  Be sure to recode genotypes in that mannyer way before passing them to this function.
	
	Marker positions in \code{R/qtl} are expressed in centimorgans, not basepairs.  On conversion,
	physical positions are faked by assuming recombination rate 1 cM per 1 Mbp and rounding to
	the next-lowest integer.

	Only crosses of type \code{"f2"} (F2 intercross) or \code{"bc"} are supported, and partially-
	informative genotypes will probably be mangled.
}
\references{
\code{R/qtl}: \url{http://www.rqtl.org}

Broman KW, Wu H, Sen S, Churchill GA. (2003) R/qtl: QTL mapping in experimental crosses.
	Bioinformatics 19:889-890. doi:10.1093/bioinformatics/btg112.

Broman KW, Sen S. (2009) A Guide to QTL Mapping with R/qtl. Springer, New York.
}
\seealso{
\code{\link[qtl]{read.cross}}, \code{\link{as.rqtl.genotypes}} (for inverse operation)
}

