% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{as.rqtl}
\alias{as.rqtl}
\title{Convert a \code{genotypes} object to an \code{R/qtl} object}
\usage{
as.rqtl(gty, type = c("f2", "bc"), chroms = paste0("chr", c(1:19, "X")),
  ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{type}{cross type for \code{R/qtl} (only \code{"bc"} [backcross] and \code{"f2"} [F2 intercross] currently supported)}

\item{chroms}{vector of chromosome names to include in output (in order)}

\item{...}{ignored}
}
\value{
an object of class \code{cross}, with the specified cross type
}
\description{
Convert a \code{genotypes} object to an \code{R/qtl} object
}
\details{
Karl Broman's \code{R/qtl} is a widely-used package for mapping quantiative traits
	in experimental crosses of laboratory organisms and crop plants.  It expects genotypes to
	be coded with respect to parental lines: eg. AA, AB, BB for an F2 cross between (true-breeding)
	lines A and B.  Be sure to recode genotypes in that mannyer way before passing them to this function.
	
	Marker positions in \code{R/qtl} are expressed in centimorgans, not basepairs, so only markers with
	non-zero, non-missing genetic positions will be included in the output of this function.
}
\references{
\code{R/qtl}: \url{http://www.rqtl.org}

Broman KW, Wu H, Sen S, Churchill GA. (2003) R/qtl: QTL mapping in experimental crosses.
	Bioinformatics 19:889-890. doi:10.1093/bioinformatics/btg112.

Broman KW, Sen S. (2009) A Guide to QTL Mapping with R/qtl. Springer, New York.
}
\seealso{
\code{\link[qtl]{read.cross}}, \code{\link{as.genotypes.cross}} (for inverse operation)
}

