% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plink.R
\name{assoc.plink}
\alias{assoc.plink}
\title{Perform a genome-wide association scan with PLINK}
\usage{
assoc.plink(prefix, model = c("assoc", "linear", "logistic"),
  test = c("additive", "genotypic", "hethom", "dominant", "recessive"),
  perms = TRUE, nperms = 1e+05, geno.missing = 0, ind.missing = 0,
  maf = 0, hwe = 1,
  flags = "--keep-allele-order --allow-no-sex --nonfounders", ...)
}
\arguments{
\item{prefix}{a pointer to a PLINK fileset (of class \code{plink})}

\item{model}{statistical model for association between genotype and phenotype (\code{"assoc"} for
case-control contingency table; \code{"linear"} for linear model on quantitative trait;
\code{"logistic"} for case-control logistic regression)}

\item{test}{family of hypothesis test(s) to perform}

\item{perms}{logical: establish significance thresholds using adaptive permutation, or skip it?}

\item{nperms}{integer; maximum number of permutations}

\item{geno.missing}{drop markers with call rate lower than this threshold}

\item{ind.missing}{drop samples with call rate lower than this threshold}

\item{maf}{drop markers with minor-allele frequency lower than this threhsold}

\item{hwe}{drop markers p-value less than this threshold for test of Hardy-Weinbery equilibrium}

\item{flags}{additional command-line flags passed directly to PLINK call}

\item{...}{ignored}
}
\value{
a dataframe with association results, having the following columns:
\itemize{
\item chr
\item pos
\item marker
\item A1 -- reference allele (interpretation depends on dataset)
\item p.value -- p-value for hypothesis test
\item OR -- odds ratio (in case-control context)
\item n -- count of non-missing genotypes at this marker
\item test -- which family of hypothesis test this is
\item p.value.perm -- empirical p-value from permutations, if applicable
}
}
\description{
Perform a genome-wide association scan with PLINK
}
\details{
See the relevant PLINK documentation for details of the underlying calculations. Note
	that PLINK doesn't perform kinship correction.  There now exists other, smarter software for
	GWAS including \code{EMMAX}, \code{FastLMM} and \code{LIMIX}.
}
\references{
PLINK v1.9: \url{https://www.cog-genomics.org/plink2}

Purcell S et al. (2007) PLINK: a toolset for whole-genome association and population-based
	linkage analysis. Am J Hum Genet 81(3): 559-575. doi:10.1086/519795.
}

