% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{autosomes}
\alias{autosomes}
\title{Shortcut for grabbing just the autosomes}
\usage{
autosomes(gty, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{...}{ignored}
}
\value{
a copy of \code{gty} without chrX, chrY or chrM (or markers not assigned to a chromosome)
}
\description{
Shortcut for grabbing just the autosomes
}

