% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{blank.grob}
\alias{blank.grob}
\title{Make a blank grob for plotting}
\usage{
blank.grob(...)
}
\arguments{
\item{...}{ignored}
}
\value{
a \code{grid::grob} object
}
\description{
Make a blank grob for plotting
}
\seealso{
\code{\link[grid]{grid.rect}}
}

