% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{$.genotypes}
\alias{$.genotypes}
\title{Access attributes of a \code{genotypes} object as if they were a list}
\usage{
\method{$}{genotypes}(x, expr, ...)
}
\arguments{
\item{x}{an object of class \code{genotypes}}

\item{expr}{name of the attribute to extract}

\item{...}{ignored}
}
\description{
Access attributes of a \code{genotypes} object as if they were a list
}
