% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geno.R
\name{cbind.genotypes}
\alias{cbind.genotypes}
\title{Concatenate two \code{genotypes} objects 'horizontally'}
\usage{
cbind.genotypes(a, b, ...)
}
\arguments{
\item{a}{an object of class \code{genotypes}}

\item{b}{another object of class \code{genotypes}}

\item{...}{ignored}
}
\description{
Concatenate two \code{genotypes} objects 'horizontally'
}
\details{
The objects \code{a} and \code{b} must represent the same markers typed on non-overlapping sets
	of samples in order for this operation to be meaningful. That means their row names must match.
	Intensity data and any sample or site filters are dropped in the result.
}

