% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics.R
\name{consensus}
\alias{consensus}
\title{Identify the consensus genotype call among a group of samples}
\usage{
consensus(gty, nas.allowed = 0, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{nas.allowed}{maximum proportion of missing genotypes to ignore}

\item{...}{ignored}
}
\value{
a named vector of consensus genotypes, with length equal to number of rows in \code{gty}
}
\description{
Identify the consensus genotype call among a group of samples
}
\details{
Define the "consensus" genotype to be the genotype with greatest frequency among a group
	of samples, with ties broken in favor of the major allele.  This function computes that consensus
	ignoring missing values, and then masks sites with missingness greater than \code{nas.allowed}.
}

