% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics.R
\name{dist.genotypes}
\alias{dist.genotypes}
\title{Calculate a simple genetic distance: proportion of alleles shared IBS}
\usage{
\method{dist}{genotypes}(gty, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{...}{ignored}
}
\value{
a \code{dist} object whose entires are the proportion of alleles shared
	identical-by-state between samples
}
\description{
Calculate a simple genetic distance: proportion of alleles shared IBS
}
\details{
To get a plain distance matrix, do \code{as.matrix(dist(x))}.
}
\seealso{
\code{\link[stats]{dist}}
}

