% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{dotplot.genotypes}
\alias{dotplot}
\alias{dotplot.genotypes}
\title{Show a visual representation of a slice of a genotype matrix}
\usage{
\method{dotplot}{genotypes}(gty, size = 2, meta = NULL, shape = c("point",
  "tile", "nothing"), color.by = NULL, force = FALSE, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{size}{numeric; size of points}

\item{meta}{dataframe of additional sample metadata to be joined to the input before plotting}

\item{shape}{if \code{"point"}, each genotype call is represented by a circle; if \code{"tile"},
the effect will be like a heatmap}

\item{color.by}{column in \code{meta} to use for color-coding points; otherwise use genotype calls}

\item{force}{logical; if \code{TRUE}, issues a stern warning before trying to make a huge
plot which might hang the \code{R} session}

\item{...}{ignored}
}
\value{
a grid of genotype calls, laid out horizontally in genomic coordinates
}
\description{
Show a visual representation of a slice of a genotype matrix
}
\details{
Each point represents a genotype call for a single sample.  Points are coloured according
	to numeric genotype call (0/1/2/N), with missing (N) calls hidden.  Points are spaced evenly along
	the x-axis to facilitate visual inspection; marker spacing in genomic coordinates is indicated in
	a track along the bottom of the plot which connects each column to its relative genomic position.
	At present, the function will only do a single chromosome at a time.

	Since the return value is a \code{ggplot}, the user can add modify it at will.  To recover the
	underlying dataframe, use \code{...$data}.
}
\seealso{
\code{\link{plot.clusters}} for intensity plots by marker
}

