% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{drop.intensity}
\alias{drop.intensity}
\title{Strip intensity matrices from a \code{genotypes} object}
\usage{
drop.intensity(gty, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{...}{ignored}
}
\value{
a copy of \code{gty} without intensity data (but with any QC summaries still present)
}
\description{
Strip intensity matrices from a \code{genotypes} object
}
