% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plink.R
\name{filter.plink}
\alias{filter.plink}
\title{Filter markers and samples from a dataset with PLINK}
\usage{
filter.plink(prefix, out = NULL, chr = NULL, from = NULL, to = NULL,
  maf = 0, hwe = 1, geno.missing = 0, ind.missing = 0, attrib = NULL,
  attrib.file = NULL, remove = NULL, keep = NULL, remove.fam = NULL,
  keep.fam = NULL, flags = "", ...)
}
\arguments{
\item{prefix}{a pointer to a PLINK fileset (of class \code{plink})}

\item{out}{filename prefix for filtered PLINK fileset; overrides \code{prefix}}

\item{chr}{keep only markers on this chromosome}

\item{from}{with \code{chr}, keep only markers with position greater than this}

\item{to}{with \code{chr}, keep only markers with position less than this}

\item{maf}{drop markers with minor-allele frequency lower than this threhsold}

\item{hwe}{drop markers p-value less than this threshold for test of Hardy-Weinbery equilibrium}

\item{geno.missing}{drop markers with call rate lower than this threshold}

\item{ind.missing}{drop samples with call rate lower than this threshold}

\item{attrib}{with \code{attrib.file}, keep only samples with these attribute(s) (labels)}

\item{attrib.file}{with \code{attrib}, a file containing family IDs, sample IDs and attribute(s)}

\item{remove}{character vector of samples to exclude}

\item{keep}{character vector of samples to keep}

\item{remove.fam}{character vector of family IDs to exclude}

\item{keep.fam}{character vector of family IDs to keep}

\item{flags}{additional command-line flags passed directly to PLINK call}

\item{...}{ignored}
}
\value{
a pointer (of class \code{plink}) to a new PLINK binary fileset after application
	of the above filters
}
\description{
Filter markers and samples from a dataset with PLINK
}
\details{
See the relevant PLINK documentation for details of the filters and precedence rules
	governing how they are applied.
}
\references{
PLINK v1.9: \url{https://www.cog-genomics.org/plink2}

Purcell S et al. (2007) PLINK: a toolset for whole-genome association and population-based
	linkage analysis. Am J Hum Genet 81(3): 559-575. doi:10.1086/519795.
}

