% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geno.R
\name{filters.genotypes}
\alias{filters}
\alias{filters.genotypes}
\title{Get filters attached to a \code{genotypes} object}
\usage{
\method{filters}{genotypes}(gty, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}
}
\value{
list of length 2 with elements \code{sites} and \code{samples} which should run parallel to the
	row and columns, respectively, of the genotypes matrix in the input.  If fitlers are absent, returns logical vectors
	with appropriate dimensions which are uniformly \code{FALSE}.  These vectors have names matching the row
	and column names of the genotypes matrix.
}
\description{
Get filters attached to a \code{genotypes} object
}
\seealso{
Other accessor functions: \code{\link{markers}} (marker map), \code{\link{samples}}
	(sample metadata), \code{\link{intensity}} (intensity matrices)
}

