% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{fix.strand.swaps}
\alias{fix.strand.swaps}
\title{Fix alleles in marker map using those in observed genotypes}
\usage{
fix.strand.swaps(gty, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}
}
\value{
A new \code{genotypes} object with corrected alleles, whose marker map has two
	additional columns: \code{swapped} indicates if alleles were swapped, and \code{unresolved}
	indicates that observed and expected alleles don't match, but can't be reconciled with
	a simple strand swap.
}
\description{
Fix alleles in marker map using those in observed genotypes
}
\details{
Because this method relies on information in observed genotypes, it is advisable
	to use it only with "many" samples, where "many" is enough to capture all alleles of interest.
}
