% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics.R
\name{fixed.diffs}
\alias{fixed.diffs}
\title{Identify fixed differences between a pair of samples}
\usage{
fixed.diffs(gty, cols = TRUE, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{cols}{indexing vector for selecting columns from \code{gty}}

\item{...}{ignored}
}
\value{
a named logical vector, of length equal to the number of rows in \code{gty},
	indicating whether a site is a fixed difference between the two samples
}
\description{
Identify fixed differences between a pair of samples
}
\details{
A site is defined to have a "fixed difference" between two samples if they have opposite
	homozygous genotypes.  In contrast to \code{segregating()}, this function rejects sites at which either
	sample is heterozygous or has a missing genotype, rather than marking them with \code{NA}s.
	
	If more than two samples are present in the input, an error occurs.
}
\seealso{
\code{\link{segregating}}
}

