% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{freqplot}
\alias{freqplot}
\title{Plot frequencies of missing, heterozygous, and minor-allele calls across genome}
\usage{
freqplot(gty, max.H = -1, max.N = -1, min.maf = Inf, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{max.H}{show markers with heterozygosity greater than this threshold}

\item{max.N}{show markers with missingness greater than this threshold}

\item{min.maf}{show markers with minor-allele frequency less than this threshold}

\item{...}{ignored}
}
\value{
a \code{ggplot} object with a Manhattan-style plot of the above values
}
\description{
Plot frequencies of missing, heterozygous, and minor-allele calls across genome
}
\seealso{
\code{\link{qcplot}}, \code{\link{bafplot}}, \code{\link{summarize.calls}}
}

