% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{genoapply}
\alias{genoapply}
\title{Apply a function over a \code{genotypes} object, by sample or marker groups}
\usage{
genoapply(gty, margin = c(1, 2), expr = 1, fn = NULL, strip = FALSE,
  ...)
}
\arguments{
\item{gty}{an object of class \code{genotypes}}

\item{margin}{dimension of the array over which to apply \code{expr} (1=rows, 2=columns)}

\item{expr}{an expression defining subsets of \code{gty}, either by markers (\code{margin = 1})
or samples (\code{margin = 1})}

\item{fn}{the function to apply over the subsets defined by \code{expr}}

\item{strip}{logical; if \code{TRUE}, strip metadata before evaluating \code{fn()}}

\item{...}{other arguments passed-through to \code{fn()}}
}
\value{
a list with one element per value of the grouping expression (\code{expr}); the value of that
	element is the return value of \code{fn()} applied to the subset of \code{gty} defined
	by \code{expr}
}
\description{
An homage to base \code{R}'s \code{apply} family of functions.
}
\details{
The grouping expression \code{expr} is evaluated in the parent environment by default. Use the
	dot function to get lazy evaluation like with \code{subset()} (see examples).

	When \code{fn()} will involve heavy use of array slicing, use \code{strip} to avoid having to
	slice sample/marker metadata repeatedly.
}
\examples{
# lazy evaluation 1: apply over chromosomes
\dontrun{
genoapply(x, 1, .(chr), myfunction, extra.arg = TRUE)
}
\dontrun{
genoapply(x, 2, .(fid), myfunction, extra.arg = TRUE)
}

}
