% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geno.R
\name{get.baf}
\alias{get.baf}
\title{Extract BAF/LRR as a dataframe}
\usage{
get.baf(gty, markers = NULL, ...)
}
\arguments{
\item{gty}{an object of class \code{genotypes}}

\item{markers}{indexing vector to choose which markers (rows) to extract}

\item{...}{ignored}
}
\value{
a dataframe containing sample metadata, marker metadata, and BAF/LRR values
}
\description{
Extract BAF/LRR as a dataframe
}
\details{
BAF and LRR must have already been computed using the \code{\link{tQN}} function.
}

