% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{get.call}
\alias{get.call}
\title{Extract genotype calls as a dataframe}
\usage{
get.call(gty, markers, ...)
}
\arguments{
\item{gty}{an object of class \code{genotypes}}

\item{markers}{indexing vector to choose which markers (rows) to extract}

\item{...}{ignored}
}
\value{
a dataframe containing sample metadata, marker metadata, and genotype calls
}
\description{
Extract genotype calls as a dataframe
}

