% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{get.intensity}
\alias{get.intensity}
\title{Extract hybridization intensities as a dataframe}
\usage{
get.intensity(gty, markers, ...)
}
\arguments{
\item{gty}{an object of class \code{genotypes}}

\item{markers}{indexing vector to choose which markers (rows) to extract}

\item{...}{ignored}
}
\value{
a dataframe containing sample metadata, marker metadata, and hybridization intensities
}
\description{
Extract hybridization intensities as a dataframe
}
