% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{ggmanhattan}
\alias{ggmanhattan}
\title{Create skeleton of a 'Manhattan plot' (concatenated chromosomes) with \code{ggplot2}}
\usage{
ggmanhattan(df, chroms = NULL, scale = c("Mbp", "cM"), space = NULL,
  cols = c("grey30", "grey60"), ...)
}
\arguments{
\item{df}{a dataframe with columns (at least) "chr" and "pos"}

\item{chroms}{chromosome names (in order); will try to guess them if not supplied}

\item{scale}{one of \code{"Mbp"} (millions of basepairs; default) or \code{"cM"} (centimorgans)}

\item{space}{padding factor to add in between chromosomes}

\item{cols}{vector of length 2, giving alternating colours for alternating chromosomes}
}
\value{
a \code{ggplot2} object whose x-axis is defined by chromosome and position (with
	chromosomes concatenated in karyotype order), to which data in \code{df} can be added as
	additional layers
}
\description{
Create skeleton of a 'Manhattan plot' (concatenated chromosomes) with \code{ggplot2}
}

