% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics.R
\name{guess.parents}
\alias{guess.parents}
\title{Attempt to guess the mother-father pair corresponding to offspring}
\usage{
guess.parents(gty, parents, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object containing target individuals}

\item{parents}{a \code{genotypes} object containing possible parents for the target individuals}

\item{...}{ignored}
}
\value{
the sample metedata from the input object, with \code{mom} and \code{dad} values replaced
	by the function's best guesses
}
\description{
Attempt to guess the mother-father pair corresponding to offspring
}
\details{
This wrapper calls \code{mendel.distance()} for all mother-father pairs available from
	\code{parents} and all offspring in \code{gty}, and for each offspring reports the mother-father
	pair with the smallest Mendel distance.  No guarantees are made with respect to near-misses.
}

