% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geno.R
\name{has.intensity}
\alias{has.intensity}
\title{Check if a \code{genotypes} object has intensity data attached}
\usage{
has.intensity(gty, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{...}{ignored}
}
\value{
logical scalar; TRUE is intensity data is present
}
\description{
Check if a \code{genotypes} object has intensity data attached
}

