% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{heatmap.genotypes}
\alias{heatmap.genotypes}
\title{Plot a heatmap representing genetic distances between samples}
\usage{
\method{heatmap}{genotypes}(gty, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}
}
\value{
a (hierarchically-clustered) matrix representation of pairwise genetic distances between
	samples, normalized to [0,1].
}
\description{
Plot a heatmap representing genetic distances between samples
}

