% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{intensity.genotypes}
\alias{intensity}
\alias{intensity.genotypes}
\title{Get intensity matrices attached to a \code{genotypes} object}
\usage{
\method{intensity}{genotypes}(gty, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{...}{ignored}
}
\value{
list of length 2 with elements \code{x} and \code{y}, respectively the x- and y-intensity matrices,
	which both have dimensions and row/column names equal to those of the genotypes matrix in the input.
}
\description{
Get intensity matrices attached to a \code{genotypes} object
}
\seealso{
Other accessor functions: \code{\link{markers}} (marker map), \code{\link{samples}}
	(sample metadata), \code{\link{filters}} (site and sample filters)
}

