% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{intensity.histogram}
\alias{intensity.histogram}
\title{Plot histogram of (sum-)intensities by sample}
\usage{
\method{intensity}{histogram}(gty, ...)
}
\description{
Plot histogram of (sum-)intensities by sample
}

