% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qc.R
\name{intensity.vs.ref}
\alias{intensity.vs.ref}
\title{KS-test for difference in intensity distributions}
\usage{
\method{intensity}{vs.ref}(gty, ref, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object with intensity data attached}

\item{ref}{a matrix, vector or object coercible to such, containing sum-intensities from reference samples}
}
\value{
a named vector of D_j, the Kolmogorov-Smirnov test statistic for each sample j
}
\description{
KS-test for difference in intensity distributions
}
\details{
This function detects potential failed arrays by performing the Kolmogorov-Smirnov test
	for difference between the "sum-intensities" (see \code{\link{summarize.intensity}}) of each sample
	and some reference distribution of "sum-intensities" of known good samples.  This test should (obviously)
	be performed *before* any normalizations are applied.  As such it may be useful for detecting batch
	effects, although that possibility has not been systematically explored.

	The distribution of "sum-intensity" across an array is expected to be approximately normal. Outliers for
	the D statistic come in two flavours (cf. Didion et al. (2014)): samples which fail completely, having a
	heavily right-skewed intensity distribution; and samples which are genetically diverged from the reference
	sample/species used in array design.  Divergent samples have a spike of intensities near zero, representing
	failed hybridization due to off-target variation, but an otherwise normal intensity distribution.
}
\references{
Didion JP et al. (2014) SNP array profiling of mouse cell lines identifies their strains of origin
	and reveals cross-contamination and widespread aneuploidy. BMC Genomics 15(1): 847. doi:10.1186/1471-2164-15-847.
}
\seealso{
\code{\link{summarize.intensity}}, \code{\link{summarize.calls}}, \code{\link{run.sample.qc}}
}

