% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{is.filtered}
\alias{is.filtered}
\title{Check if markers or samples are marked with filters}
\usage{
is.filtered(gty, hits = 0, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{hits}{integer; maximum number of filters which can be set before a marker or sample is flagged}

\item{...}{ignored}
}
\value{
a list with two elements: \code{$sites}, logical vector of filter status for markers;
	and \code{$samples}, logical vector of filter status for samples
}
\description{
Check if markers or samples are marked with filters
}

