% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{kinship.plink}
\alias{kinship.plink}
\title{Compute realized kinship matrix (aka GRM) with PLINK}
\usage{
kinship.plink(prefix, method = "--distance square 1-ibs", flags = "",
  prune = FALSE, ...)
}
\arguments{
\item{prefix}{a pointer to a PLINK fileset (of class \code{plink})}

\item{method}{command-line flags for \code{PLINK}, controlling underlying calculation}

\item{flags}{command-line flags passed directly to underlying PLINK call}

\item{prune}{logical; if \code{TRUE}, do a round of LD-pruning before calculating kinship matrix}

\item{...}{ignored}
}
\value{
a square matrix of genetic distances (or similarities), with row and column names
	equal to the sample IDs
}
\description{
Compute realized kinship matrix (aka GRM) with PLINK
}
\details{
See the relevant PLINK documentation for details of the underlying calculations.
	The default is to compute the 'Mahnattan distance' between samples, ie. 1 minus the mean
	number alleles shared identical-by-state across all markers.
}
\references{
PLINK v1.9: \url{https://www.cog-genomics.org/plink2}

Purcell S et al. (2007) PLINK: a toolset for whole-genome association and population-based 
	linkage analysis. Am J Hum Genet 81(3): 559-575. doi:10.1086/519795.
}

