% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plink.R
\name{ld.plink}
\alias{ld.plink}
\title{Compute pairwise LD between markers with PLINK}
\usage{
ld.plink(prefix, dprime = FALSE, index.snp = NULL, markers = NULL,
  chr = NULL, from = NULL, to = NULL, window = NULL, window.r2 = 0,
  flags = "", ...)
}
\arguments{
\item{prefix}{a pointer to a PLINK fileset (of class \code{plink})}

\item{dprime}{logical; if \code{TRUE}, compute Lewontin D-prime instead of R2}

\item{index.snp}{name of the marker used to anchor the calculation}

\item{markers}{compute all pairwise LD values between markers in this list}

\item{chr}{limit analysis to this chromosome}

\item{from}{with \code{chr}, limit analysis to window starting at this position}

\item{to}{with \code{chr}, limit analysis to window ending at this position}

\item{window}{(ignored)}

\item{window.r2}{keep calculating until pairwise LD falls below this threshold}

\item{flags}{command-line flags passed directly to underlying PLINK call}
}
\value{
a \code{data.table} of pairwise LD values, one marker pair per row
}
\description{
Compute pairwise LD between markers with PLINK
}
\details{
Linkage disequlibrium (LD) is estimated here via the squared genetic correlation (r^2)
	between loci.  Since the size of the output scales with the square of the number of markers and
	can easily grow to gigabytes for a dataset with >10k markers, it is highly recommended
	to limit the calculation using the parameters above.

	See the relevant PLINK documentation for details of the underlying calculations.
}
\references{
PLINK v1.9: \url{https://www.cog-genomics.org/plink2}

Purcell S et al. (2007) PLINK: a toolset for whole-genome association and population-based
	linkage analysis. Am J Hum Genet 81(3): 559-575. doi:10.1086/519795.

Lewontin RC (1960) The interaction of selection and linkage. I. General considerations; heterotic
	models. Genetics 49(1): 49-67.
}

