% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics.R
\name{maf}
\alias{maf}
\title{Calculate minor-allele frequency (MAFs) from a genotypes matrix}
\usage{
maf(...)
}
\arguments{
\item{...}{passed through to call to \code{freq()}}
}
\description{
Just an alias or \code{freq(x, "markers")}.  Sticklers for nomenclature will be aware that this
function does not truly return the MAF unless the allele encoding is \code{"relative"}.
}
\seealso{
\code{\link{freq}}
}

