% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geno.R
\name{markers.genotypes}
\alias{markers.genotypes}
\title{Get marker map for a \code{genotypes} object}
\usage{
\method{markers}{genotypes}(gty, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}
}
\value{
a dataframe of marker information (chr, cM, pos, ...) which should run parallel to the
	rows of the genotypes matrix in the input.  If map is absent, returns \code{NULL.}
}
\description{
Get marker map for a \code{genotypes} object
}
\seealso{
Other accessor functions: \code{\link{samples}} (sample metadata), \code{\link{filters}}
	(site and sample filters), \code{\link{intensity}} (intensity matrices)
}

