% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plink.R
\name{mds.plink}
\alias{mds.plink}
\title{Perform classical multidimensional scaling (MDS) with PLINK}
\usage{
mds.plink(prefix, flags = "--autosome", K = 3, ...)
}
\arguments{
\item{prefix}{a pointer to a PLINK fileset (of class \code{plink})}

\item{flags}{command-line flags passed directly to underlying PLINK call}

\item{K}{project samples onto this many dimensions}

\item{...}{ignored}
}
\value{
a dataframe with individual IDs, family IDs, and then projections in columns
	"MDS1"..."MDS{k}".
}
\description{
Perform classical multidimensional scaling (MDS) with PLINK
}
\details{
See the relevant PLINK documentation for details of the underlying calculations.
	The default is to perform MDS on autosomal genotypes only.  Scaled eigenvalues (ie. percent
	variance explained by each dimension) are provided in \code{attr(,"eigvals")}.
}
\references{
PLINK v1.9: \url{https://www.cog-genomics.org/plink2}

Purcell S et al. (2007) PLINK: a toolset for whole-genome association and population-based
	linkage analysis. Am J Hum Genet 81(3): 559-575. doi:10.1086/519795.
}
\seealso{
\code{\link{pca.plink}}
}

