% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics.R
\name{mendel.distance}
\alias{mendel.distance}
\title{Compute rate of Mendelian inconsistency between an individual and possible mother-father pairs.}
\usage{
mendel.distance(gty, parents, verbose = TRUE, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object containing exactly one target individual}

\item{parents}{a \code{genotypes} object containing possible parents for the target individual}

\item{verbose}{logical; if \code{TRUE}, send progress messages to terminal}

\item{...}{ignored}
}
\value{
a dataframe of mother-father pairs and corresponding Mendel distances
}
\description{
Compute rate of Mendelian inconsistency between an individual and possible mother-father pairs.
}
\details{
Define the "Mendel distance" between an individual and two possible parents as the mean
	rate of Mendelian inconsistencies -- that is, markers at which zero alleles are shared IBS between
	one parent and one offspring -- across the pairs (father, offspring) and (mother, offspring). Given
	a set of possible parents (in \code{parents}), this function creates all possible male-female pairs
	and computes the Mendel distance between each pair and a single target individual (in \code{gty}).
	
	If the sex of the target individual is non-missing, the sex chromosomes will be handled specially:
	for males, chrY is compared to the father's chrY and chrX to the mother's chrX; for females, chrX
	is compared to both parents and chrY is ignored.  If the sex of the target individual is missing,
	the sex chromosomes will be ignored, and mitochondrial markers are always ignored because they are
	so few in number and tend to be homoplasic.
}
\seealso{
\code{\link{guess.parents}}
}
