% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genetics.R
\name{missingness}
\alias{missingness}
\title{Calculate rate of missing genoypes by marker or sample}
\usage{
missingness(gty, by = c("markers", "samples"), ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{by}{margin over which to count missing calls}

\item{...}{ignored}
}
\value{
a named vector of missing-genotype *proportions*
}
\description{
Calculate rate of missing genoypes by marker or sample
}

