% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genetics.R
\name{nmiss}
\alias{nmiss}
\title{Count number of missing genotypes by marker or sample}
\usage{
nmiss(gty, by = c("markers", "samples"), ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{by}{margin over which to count missing calls}

\item{...}{ignored}
}
\value{
a named vector of missing-genotype *counts*
}
\description{
Count number of missing genotypes by marker or sample
}

