% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pheno.R
\name{oneway.plot}
\alias{oneway.plot}
\title{Make graphical representation of a 'one-way' genotype-phenotype contingency table}
\usage{
oneway.plot(df, space = 1, ...)
}
\arguments{
\item{df}{a dataframe from \code{oneway()}}

\item{space}{numeric scalar; buffer to add between adjacent squares in the plot}

\item{...}{ignored}
}
\value{
a \code{ggplot} object equivalent to output from base-\code{R}'s \code{mosaicplot()},
	plotting the (genotype x phenotype) contingency table for a single marker
}
\description{
Make graphical representation of a 'one-way' genotype-phenotype contingency table
}
\details{
A two-way contingency table can be rendered as a tesselation of rectangles whose area
	is proportional to the corresponding cell counts.  (This is sometimes called a "fluctiation plot".)
	In this version the plot is divided into columns by genotype, and the width of those columns
	is proportional to genotype frequency.

	Due to low-level manipulation of axis breaks and labels, this function can only handle
	a single marker at a time.  A warning will be issued if the input contains data from multiple
	markerks, although a (probably bogus) plot will still be drawn.
}
\seealso{
\code{\link{oneway}}
}

