% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pheno.R
\name{oneway}
\alias{oneway}
\title{Get one-way marker-phenotype association}
\usage{
oneway(gty, marker, verbose = TRUE, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{marker}{indexing vector into \code{gty}: can be logical, character or numeric}
}
\value{
a dataframe with marker metadata, genotype calls, and sample metadata including phenotype
}
\description{
Get one-way marker-phenotype association
}
\details{
To facilitate plotting downstream, the function tries to guess if the phenotype is binary
	(assuming the \code{PLINK} convention of 1=control/2=case), ordinal/categorical, or continuous.
	Categorical phenotypes will be converted to a factor.
}

