% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.genotypes}
\alias{pca.genotypes}
\title{Perform PCA on a \code{genotypes} object}
\usage{
\method{pca}{genotypes}(x, extras = NULL, what = c("genotypes",
  "intensity"), K = 3, fast = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{genotypes}}

\item{extras}{a second dataset of class \code{genotypes}, to be projected onto PCs computed
from the first. NB: this function *does not* verify that the underlying features (ie. markers) match.}

\item{what}{\code{"genotypes"} to do PCA on genotypes (coded 0/1/2); \code{"intensity"} to do PCA on
underlying 2D intensities.  The latter triggers an error if intensity matrices are absent.}

\item{K}{how many PCs to return.}

\item{fast}{if \code{TRUE}, use \code{corpcor::fast.svd()} to speed up calculations}

\item{...}{other parameters for call to \code{prcomp}, such as \code{center} and \code{scale}
    (both \code{TRUE} by default)}
}
\value{
a dataframe with as many rows as samples, in which the first columns are sample IDs and any associated
	metadata (as returned by \code{samples(x)}), followed by the first K PCs.  Scaled eigenvalues (ie. percent
	of variance explained) are provided as \code{attr(,"explained")}.
}
\description{
Performs principal components analysis (PCA) on either (numerically-coded) genotypes or
on the underliny 2D hybridization-intensity matrices.
}
\details{
Uses base-\code{R}'s \code{prcomp} under the hood (unless \code{fast = TRUE}). By default,
	columns are centered and scaled; not doing so will likely produce a strange result.  When doing PCA on genotypes,
	missing values are replaced with the column mean, which in many circumstances can be interpreted as the
	minor-allele frequency. (This is very similar to the behavior of PLINK.)  When doing PCA on intensities, missing
	values are set to zero -- but even no-call genotypes have nonmissing intensities on Illumina arrays, so this is unlikely
	to have any effect in practice.
}
\seealso{
\code{pca.plink()} for using PLINK's (much faster and more powerful) implementation
}

