% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plink.R
\name{pca.plink}
\alias{pca.plink}
\title{Perform PCA on genotypes with PLINK}
\usage{
\method{pca}{plink}(prefix, flags = "--autosome", K = 20, by = c("indiv",
  "var"), ...)
}
\arguments{
\item{prefix}{a pointer to a PLINK fileset (of class \code{plink})}

\item{flags}{command-line flags passed directly to underlying PLINK call}

\item{K}{return the projection of samples onto the top K PCs}

\item{by}{project individuals (\code{"indiv"}) or markers (\code{"var"}) onto PCs?}
}
\value{
When \code{by = "indiv"} (the default), a dataframe with individual IDs, family IDs,
and then projections in columns "PC1"..."PC{k}".  When \code{by == "var"}, a dataframe with
columns chromosome and marker name followed by PCs.
}
\description{
Perform PCA on genotypes with PLINK
}
\details{
See the relevant PLINK documentation for details of the underlying calculations.
	The default is to perform PCA on autosomal genotypes only.  Scaled eigenvalues (ie. percent
	variance explained by each dimension) are provided in \code{attr(,"eigvals")}.

	A recently-proposed test for natural selection uses the squared loadings of each marker
	against the top PCs as an analog of F_st.  To obtain those loadings use \code{by = "var"}.
}
\references{
PLINK v1.9: \url{https://www.cog-genomics.org/plink2}

Purcell S et al. (2007) PLINK: a toolset for whole-genome association and population-based
	linkage analysis. Am J Hum Genet 81(3): 559-575. doi:10.1086/519795.

Duforet-Frebourg N et al. (2015) Detecting genomic signatures of natural selection with
	principal componentanalysis: application to the 1000 Genomes data
	\url{http://arxiv.org/abs/1504.04543}
}
\seealso{
\code{\link{mds.plink}}
}

