% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plink.R
\name{plinkify}
\alias{plinkify}
\title{Create a pointer to a PLINK fileset}
\usage{
plinkify(gty, where = tempdir(), prefix = "stuff", flags = "", ...)
}
\arguments{
\item{gty}{a length-1 character vector containing the path to an existing PLINK fileset, OR a
\code{genotypes} object which will be converted to one}

\item{where}{if \code{gty}, directory in which to write the new PLINK fileset}

\item{prefix}{basename of the fileset to be created}

\item{flags}{(ignored)}

\item{map}{if \code{gty} a \code{genotypes} object, a marker map to override its current one}

\item{map}{if \code{gty} a \code{genotypes} object, sample metadata to override any existing metadata}
}
\value{
a length-1 character vector with path to a PLINK binary fileset, of class \code{"plink"}
}
\description{
Create a pointer to a PLINK fileset
}
\details{
Command-line PLINK takes its input from filesets on disk. This function creates a pointer to such
	a fileset after checking that all its members (\code{*.bed}, \code{*.bim}, \code{*.fam}) are present.
	Wrapper functions for various PLINK analyses accept this pointer as input.

	When the fileset alraedy exists, an additional attribute (\code{"where"}) is added to the result
	specifying where to place the output of any future PLINK calls.  By default such files are routed
	to the temporary directory corresponding to the current \code{R} session, to avoid cluttering the
	directory containing the source fileset.

	When the fileset does not exist and \code{gty} is a \code{genotypes} object, a new fileset is created
	in the location specified by \code{where} via a call to \code{write.plink(gty)}.  Then the path to
	that fileset is returned.
}

