% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.clusters}
\alias{plot.clusters}
\title{Plot 2D hybridization intensities at a few markers}
\usage{
plot.clusters(x, markers = NULL, theme.fn = ggplot2::theme_bw,
  force = FALSE, ...)
}
\arguments{
\item{x}{a \code{genotypes} object}

\item{markers}{indexing vector for selecting markers from \code{x}}

\item{theme.fn}{a function specifying formatting options for the resulting \code{ggplot}}

\item{force}{logical; if \code{TRUE}, issues a stern warning before trying to make a huge
plot which might hang the \code{R} session}

\item{...}{ignored}
}
\value{
2D intensity plots, one panel per marker (via \code{ggplot})
}
\description{
Plot 2D hybridization intensities at a few markers
}
\details{
Each point represents a genotype call for a single sample.  Points are coloured according
	to numeric genotype call (0/1/2/N), with missing (N) calls in grey.  Samples for which the
	quality filter is set are shown as open circles.

	Since the return value is a \code{ggplot}, the user can add modify it at will.  To recover the
	underlying dataframe, use \code{...$data}.
}
\seealso{
\code{\link{dotplot.genotypes}} for compact plotting of genotype calls only
}

