% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.haplotypes}
\alias{plot.haplotypes}
\title{Plot the results of haplotype reconstruction}
\usage{
plot.haplotypes(x, space = 0.1, ...)
}
\arguments{
\item{x}{a dataframe containing the results of \code{reconstruct.haps()}}

\item{space}{numeric scalar; how much buffer to add between adjacent chromosomes}

\item{...}{ignored}
}
\value{
A \code{ggplot} object with haplotype mosaics arranged in a grid layout
}
\description{
Plot the results of haplotype reconstruction
}
\details{
The input dataframe is assumed to have the following columns: \code{.id}, \code{chr},
\code{start}, \code{end}, \code{hap1} and \code{hap2}.  Since the result is a \code{ggplot},
it can be futher decorated/modified at will.
}

