% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.pca.result}
\alias{plot.pca.result}
\title{Auto-plotting of a PCA result}
\usage{
plot.pca.result(pc, K = c(1, 2), screeplot = FALSE, show = c("points",
  "labels"), theme.fn = ggplot2::theme_bw, ...)
}
\arguments{
\item{pc}{a \code{pca.result} object}

\item{K}{numeric vector of length 2 specifing which PCs to plot against each other; first is on x-axis and second on y-axis}

\item{screeplot}{logical; if \code{TRUE}, show both the plot of two PCs against each other and the variances explained of
all available PCs}

\item{show}{length-1 character vector; show just points, or sample IDs}

\item{theme.fn}{a \code{ggplot2}-compatible function to specify formatting}
}
\value{
if \code{screeplot = FALSE}, a \code{ggplot}; if \code{screeplot = TRUE}, a \code{gtable::gtable} object which can be
	modified or re-rendered with \code{grid::grid.draw()}.
}
\description{
Auto-plotting of a PCA result
}

