% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.haplogroup}
\alias{predict.haplogroup}
\title{Predict haplogroup (eg chrY or chrM), given labelled controls}
\usage{
\method{predict}{haplogroup}(gty, train, method = c("lda", "kmeans"),
  what = c("intensity", "genotypes"), noise = 0.1, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object with just the markers of interest}

\item{train}{a \code{genotypes} object with labelled training samples}
}
\value{
a dataframe with predicted haplogroup membership for training samples;
	the result from the LDA procedure (see Details) is attached as \code{attr(,"lda")}.
}
\description{
Predict haplogroup (eg chrY or chrM), given labelled controls
}
\details{
We use linear discriminant analysis via (\code{MASS::lda()}) to extract the most
	informative dimensions of the input data, given the labelled training samples, and then
	project the test samples onto that dimension.
}
