% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geno.R
\name{print.genotypes}
\alias{print.genotypes}
\title{Show a concise summary of a \code{genotypes} object}
\usage{
\method{print}{genotypes}(gty, ...)
}
\arguments{
\item{gty}{an object of class \code{genotypes}}

\item{...}{ignored}
}
\description{
Show a concise summary of a \code{genotypes} object
}
\details{
This method exists solely to avoid accidentally flooding the terminal with the
	contents of a genotypes object.
}

