% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plink.R
\name{print.plink}
\alias{print.plink}
\title{Show quick summary of a \code{PLINK} pointer}
\usage{
\method{print}{plink}(prefix, ...)
}
\arguments{
\item{prefix}{a pointer to a \code{PLINK} fileset on disk (class \code{plink})}

\item{...}{ignored}
}
\description{
Show quick summary of a \code{PLINK} pointer
}

