% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genetics.R
\name{prop.het}
\alias{prop.het}
\title{Calculate the proportion of sites at which each sample is heterozygous}
\usage{
prop.het(gty, na.rm = TRUE, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{na.rm}{logical; ignore missing genotypes when counting heterozygous sites?}
}
\value{
a named vector with the proportion of sites of heterozygous for each sample
}
\description{
Calculate the proportion of sites at which each sample is heterozygous
}

