% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{prune.plink}
\alias{prune.plink}
\title{Prune markers by pairwise LD with PLINK}
\usage{
prune.plink(prefix, prune.by = "--make-founders --indep 50 5 2", ...)
}
\arguments{
\item{prefix}{a pointer to a PLINK fileset (of class \code{plink})}

\item{prune.by}{command-line flags to control LD-pruning}

\item{...}{ignored}
}
\value{
a pointer to a new, pruned PLINK fileset
}
\description{
Prune markers by pairwise LD with PLINK
}
\details{
See the relevant PLINK documentation for details of the underlying calculations.
}
\references{
PLINK v1.9: \url{https://www.cog-genomics.org/plink2}

Purcell S et al. (2007) PLINK: a toolset for whole-genome association and population-based 
	linkage analysis. Am J Hum Genet 81(3): 559-575. doi:10.1086/519795.
}

