% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{qcplot}
\alias{qcplot}
\title{Produce a visual summary of QC measures}
\usage{
qcplot(gty, draw = TRUE, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object with intensity data attached}

\item{draw}{actually show the plot, in addition to returning it}

\item{...}{ignored}
}
\value{
a \code{gtable} (ineriting from \code{grid::grob}) containg the composite QC plot (see Details)
}
\description{
Produce a visual summary of QC measures
}
\details{
This function will plot any existing QC result attached to \code{gty}; if none is present,
	\code{run.sample.qc(gty)} will be run to generate it.

	The QC plot has two panels. The upper panel displays the distribution of A, B, H and N (missing) calls for
	each sample.  The lower panel displays a the "sum intensity" quantiles of each sample.
	Samples are sorted from left to right in increasing order of median intensity, and the sort order is matched
	between panels.  Samples for which the quality filter is set are marked with an open dot in the upper panel.

	Although somewhat crude, the count of H and N calls relative to expectations is anecdotally a robust
	measure of genotyping quality (see Didion et al. (2014)).  (But the expectations are important: an outbred
	sample and an inbred sample should have very different numbers of Hs, but probably similar number of Ns.)
	Higher variance in the hybridization intensities, as indicated by wider spread of the intensity quantiles in
	the lower plot, suggests poor input DNA quality.  Contamination of one sample with another, if both had good
	quality DNA, will increase the proportion of no-calls but probably will not shift the intensity quantiles much.

	Samples which are diverged from the reference genome used in the array design are expected to have a
	higher proportion of no-calls due to off-target variation in or near the probe sequence.  See Didion
	et al. (2012) for a fuller discussion.

	For discussion of the Illumina Infinium chemistry, see Steemers et al. (2006); for more on intensity QC,
	see Staaf et al. (2008).
}
\references{
Steemers FJ et al. (2006) Whole-genome genotyping with the single-base extension assay. Nat Methods 3:31-33.
	doi:10.1038/nmeth842.

Staaf J et al. (2008) BMC Bioinformatics. doi:10.1186/1471-2105-9-409.

Didion JP et al. (2012) Discovery of novel variants in genotyping arrays improves genotype retention and
	reduces ascertainment bias. BMC Genomics 13:34. doi:10.1186/1471-2164-13-34.

Didion JP et al. (2014) SNP array profiling of mouse cell lines identifies their strains of origin and reveals
	cross-contamination and widespread aneuploidy. BMC Genomics 15:847. doi:10.1186/1471-2164-15-847.
}

