% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalize.R
\name{quantile.normalize}
\alias{quantile.normalize}
\title{Perform quantile normalization of intensity data.}
\usage{
quantile.normalize(x, weights = NULL, force = FALSE, ...)
}
\arguments{
\item{x}{a \code{genotypes} object}

\item{weights}{a vector of column weights}

\item{force}{re-run the normalization procedure even if \code{attr(,"normalized")} is set to \code{TRUE}}

\item{...}{ignored}
}
\value{
A copy of the input object, with raw intensities replaced by the normalized ones.
}
\description{
Perform quantile normalization of intensity data.
}
\details{
A simple wrapper around the quantile normalization originally described in Bolstad et al. (2003),
	as implemented in \code{preprocessCore::normalize.quantiles.robust()}.
}
\references{
Bolstad BM et al. (2003) A comparison of normalization methods for high density oligonucleotide
	array data based on bias and variance. Bioinformatics 19(2): 185-193.
}

