% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geno.R
\name{rbind.genotypes}
\alias{rbind.genotypes}
\title{Concatenate two \code{genotypes} objects 'vertically'}
\usage{
rbind.genotypes(a, b, ...)
}
\arguments{
\item{a}{an object of class \code{genotypes}}

\item{b}{another object of class \code{genotypes}}

\item{...}{ignored}
}
\description{
Concatenate two \code{genotypes} objects 'vertically'
}
\details{
The objects \code{a} and \code{b} must represent the same samples at non-overlapping sets
	of markers in order for this operation to be meaningful. That means their column names must match.
	Sample or site filters are dropped in the result.
}

