% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{read.beadstudio}
\alias{read.beadstudio}
\title{Read genotype calls and hybridization from Illumina BeadStudio output.}
\usage{
read.beadstudio(prefix, snps, in.path = ".", keep.intensity = TRUE, ...)
}
\arguments{
\item{prefix}{filename prefix, without working directory: the \code{*} in \code{*_FinalReport.zip}}

\item{snps}{dataframe containing marker map for this array, in PLINK's \code{*.bim} format
(chromosome, marker name, cM position, bp position); rownames should be set to marker names,
and those names should match those in the BeadStudio output.}

\item{in.path}{directory in which to search for input files}

\item{keep.intensity}{should hybridization intensities be kept in addition to genotype calls?}
}
\value{
A \code{genotypes} object with genotype calls, marker map, sample metadata and (as requested)
	intensity data.
}
\description{
Read genotype calls and hybridization from Illumina BeadStudio output.
}
\details{
This function initializes a \code{genotypes} object from Illumina BeadStudio output. (For an
	example of the format, see the files in this package's \code{data/} directory.)  The two relevant
	files are \code{Sample_Map.zip} and \code{*FinalReport.zip}, which contain the sample manifest
	and genotype/intensity data, respectively.  On platforms with \code{unzip} available on the
	command line, files will be unzipped on the fly.  Otherwise \code{FinalReport.zip} (but not
	\code{Sample_Map.zip}) must be unzipped first.  This is due to the use of \code{data.table} to
	handle the usually very large genotypes file.

	The behavior of this function with respect to missing data in the genotypes versus the contents
	of \code{snps} is asymmetric.  Markers in \code{snps} which are absent in the input files will
	be present in the output, but with missing calls and intensities.  Markers in the input files
	which are missing from \code{snps} will simply be dropped.  If that occurs, check that the marker
	names in \code{snps} match exactly those in the input file.
}
\references{
Inspiration from Dan Gatti's DOQTL package: <https://github.com/dmgatti/DOQTL/blob/master/R/extract.raw.data.R>
}

