% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{recode.to.parent}
\alias{recode.to.parent}
\title{Recode genotypes against genotypes of a parent}
\usage{
\method{recode}{to.parent}(gty, parent, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{parent}{a numeric vector of parental genotypes; a scalar pointing to a reference sample in
the input (\code{gty}); or another \code{genotypes} with parental genotypes in the first column}

\item{...}{ignored}
}
\value{
a recoded \code{genotypes} object
}
\description{
Recode genotypes against genotypes of a parent
}
