% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{reconstruct.haps}
\alias{reconstruct.haps}
\title{Perform (simplistic) haplotype reconstruction, given parental genotypes}
\usage{
reconstruct.haps(geno, parents, scale = 10, err = 0.001, ...)
}
\arguments{
\item{geno}{a \code{genotypes} object}

\item{parents}{a matrix of parental genotypes}

\item{scale}{scaling parameter for expected recombination fraciton (see Details)}

\item{err}{genotyping error rate}

\item{...}{ignored}
}
\value{
a dataframe of haplotype segments
}
\description{
Perform (simplistic) haplotype reconstruction, given parental genotypes
}
\details{
This function attempts to reconstruct haplotypes of target individuals (in \code{geno})
	in terms of the founder genotypes given in \code{parents}.  It expects founders to be inbred;
	markers at which any founder has a heterozygous genotype are dropped.  (TODO: relax that
	assumption.)
	
	Transition probabilities in the HMM are proportional to expected recombination fractions,
	calculated with the Haldane map function from genetic positions of markers.  The \code{scale}
	parameter is a multiplier for these fractions when the target individuals are >1 generation
	removed from the founders.  Tuning the transition probabilities will be a matter of trial and
	error, and optimal values will probably vary by dataset.
	
	Haplotype segments are pseudo-phased in the result: phase is set to greedily minimize the number of
	haplotype transitions between adjacent segments.  (There is no look-ahead, so in some cases the
	result may be less than optimal.)  Real phasing is beyond the scope of this package at present.

 The model is constrained to only the homozygous states when a sample should really be hemizygous:
 this is true for males on chrX and everybody on chrM.
}

