% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geno.R
\name{replace.map}
\alias{replace.map}
\title{Swap out the marker map for a different one}
\usage{
replace.map(gty, newmap, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{newmap}{a valid marker map (especially: rownames same as marker names)}

\item{...}{ignored}
}
\value{
a copy of \code{gty} with old marker map replaced by the new one
}
\description{
Swap out the marker map for a different one
}
\details{
Replacement of the marker map relies on marker names: only markers with which are shared
	by the old and new maps are retained in the output.  (As such, this function is also a backdoor
	subsetting function.)  Intensity matrices, BAF/LRR matrices, and site filters are adjusted accordingly.
	A validity check is performed before the function returns to flag problems.
}

