% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geno.R
\name{replace.names}
\alias{replace.names}
\title{Rename some or all samples}
\usage{
replace.names(gty, nn, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{nn}{a vector of new sample names}

\item{...}{ignored}
}
\description{
Rename some or all samples
}
\details{
The values in \code{nn} are used to replace existing names using the following rules. First, if \code{nn}
	is a named vector, it is assumed to map old names to new names.  Otherwise \code{nn} must have length equal to the
	number of samples in \code{gty}, and each element should be a new name to replace an old one.  Missing values are
	not permitted, for obvious reasons.
}

