% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qc.R
\name{run.marker.qc}
\alias{run.marker.qc}
\title{Perform basic marker-wise QC on genotype calls}
\usage{
run.marker.qc(gty, max.H = Inf, max.N = Inf, min.hom = 0, hits = 0,
  apply = FALSE, overwrite = TRUE, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{max.H}{threshold for count of heterozygous calls, above which a site is flagged}

\item{max.N}{threshold for count of no-calls, above which a site is flagged}

\item{min.hom}{threshold for count of homozygous calls, at or below which a site is flagged}

\item{hits}{markers failing more than this many filters are flagged}

\item{apply}{logical; if \code{TRUE}, remove markers failing the filters, rather than flagging them}

\item{overwrite}{logical; if \code{TRUE} (the default), replace exisitng filters with the results of this QC check}

\item{...}{ignored}
}
\value{
a copy of the input with sample filters set, and an object of class \code{marker.QC.result} in
	attr(,"marker.qc")
}
\description{
Perform basic marker-wise QC on genotype calls
}
\details{
A wrapper for the marker-level QC functions, applied to genotype calls.  Samplies which
	fail are flagged but not actually dropped from the result (unless \code{apply = TRUE}.)
}
\seealso{
\code{\link{summarize.calls}}, \code{\link{run.sample.qc}}, \code{\link{apply.filters}}
}

