% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{run.sample.qc}
\alias{run.sample.qc}
\title{Perform basic sample-wise QC on genotype calls and intensities}
\usage{
run.sample.qc(gty, ref.intensity = NULL, max.H = Inf, max.N = Inf,
  max.D = Inf, min.D = Inf, apply = FALSE, hits = 0, overwrite = TRUE,
  ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{ref.intensity}{a vector of "sum-intensities" from known good reference samples}

\item{max.H}{threshold for count of heterozygous calls, above which a sample is flagged;
OR a named list of thresholds, with names to match "family ID" (column \code{fid}) in pedigree}

\item{max.N}{threshold for count of no-calls, above which a sample is flagged; OR a named list
as above}

\item{max.D}{upper threshold for D-statistic (see \code{\link{intensity.vs.ref}}) above which a
sample is flagged; OR a named list as above}

\item{min.D}{lower threshold for D-statistic (see \code{\link{intensity.vs.ref}}) above which a
sample is flagged; OR a named list as above}

\item{apply}{logical; if \code{TRUE}, remove samples failing the filters, rather than flagging them}

\item{hits}{samples failing more than this many filters are flagged}

\item{overwrite}{logical; if \code{TRUE} (the default), replace exisitng filters with the results of this QC check}

\item{...}{ignored}
}
\value{
a copy of the input with sample filters set, and an object of class \code{QC.result} in
	attr(,"qc")
}
\description{
Perform basic sample-wise QC on genotype calls and intensities
}
\details{
A wrapper for the sample-level QC functions, applied to genotype calls (always) and
	hybridization intensities (if present.)  Samplies which fail are flagged but not actually
	dropped from the result.
}
\seealso{
\code{\link{summarize.calls}}, \code{\link{intensity.vs.ref}}, \code{\link{apply.filters}}
}

