% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{samples.genotypes}
\alias{samples.genotypes}
\alias{samples}
\title{Get sample metadata for a \code{genotypes} object}
\usage{
\method{samples}{genotypes}(gty, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{...}{ignored}
}
\value{
a dataframe of sample metadata (iid, ...) which should run parallel to the
	columns of the genotypes matrix in the input.  If sample data is absent, returns \code{NULL.}
}
\description{
Get sample metadata for a \code{genotypes} object
}
\seealso{
Other accessor functions: \code{\link{markers}} (marker map), \code{\link{filters}}
	(site and sample filters), \code{\link{intensity}} (intensity matrices)
}
