% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_fill_heatmap}
\alias{scale_fill_heatmap}
\title{A red-to-orange color scale for heatmaps}
\usage{
scale_fill_heatmap(...)
}
\arguments{
\item{...}{additional arguments passed to \code{ggplot2::scale_fill_gradientn()}}
}
\value{
a \code{ggplot2} scale function
}
\description{
A red-to-orange color scale for heatmaps
}

