% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_x_genome}
\alias{scale_x_genome}
\title{Custom x-axis scale for genome coordinates.}
\usage{
scale_x_genome(scale = 1000000, unit = NULL, ...)
}
\arguments{
\item{scale}{scale factor, in base pairs}

\item{unit}{abbreviation for units corresponding to scale factor (see Details)}

\item{...}{additional arguments to \code{ggplot2::scale_x_continuous()}}
}
\value{
a \code{ggplot2} scale function
}
\description{
Custom x-axis scale for genome coordinates.
}
\details{
For example, if \code{scale = 1000}, we might set \code{unit = "kbp"}.
}
\seealso{
\code{\link[ggplot2]{scale_x_continuous}}
}
