% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_y_logp}
\alias{scale_y_logp}
\title{Custom y-axis scale for -log10(p-values)}
\usage{
scale_y_logp(...)
}
\arguments{
\item{...}{additional arguments to \code{ggplot2::scale_y_continuous()}}
}
\value{
a \code{ggplot2} scale function
}
\description{
Custom y-axis scale for -log10(p-values)
}
\seealso{
\code{\link[ggplot2]{scale_x_continuous}}
}

