% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics.R
\name{segregating}
\alias{segregating}
\title{Identify segregating sites among a group of samples}
\usage{
segregating(gty, nas.allowed = 0, allow.het = TRUE, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{nas.allowed}{maximum proportion of missing genotypes to ignore}

\item{allow.het}{logical; if \code{FALSE}, ignore samples with heterozygous genotype}

\item{...}{ignored}
}
\value{
a named logical vector, of length equal to the number of rows in \code{gty},
	indicating whether a site is segregating
}
\description{
Identify segregating sites among a group of samples
}
\details{
A site is defined to be "segregating" if two or more alleles are present in one or
	more copies across all samples.  Sites at which missingness is greater than \code{nas.missing}
	are set to \code{NA}.
	
	When \code{allow.het = FALSE}, this is equivalent to asking whether at least one pair of samples
	has a fixed difference at each site.
}
\seealso{
\code{\link{fixed.diffs}}
}
