% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{sex.genotypes}
\alias{sex.genotypes}
\title{Extract vector of sample sexes}
\usage{
\method{sex}{genotypes}(gty)
}
\arguments{
\item{gty}{an object of class \code{genotypes}}
}
\value{
named numeric vector giving sample sexes: 0=unknown, 1=male, 2=female
}
\description{
Extract vector of sample sexes
}

