% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{sextable}
\alias{sextable}
\title{Print summary table of sample sexes}
\usage{
sextable(gty, ...)
}
\arguments{
\item{gty}{an object of class \code{genotypes}}

\item{...}{ignored}
}
\value{
table of sample counts by sex
}
\description{
Counts the number of samples of each sex, converts numeric codes to human-friendly
labels, and returns the table.
}

