% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex.R
\name{sexysum}
\alias{sexysum}
\title{Summary stats for the sex chromosomes}
\usage{
sexysum(gty, clean = TRUE, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{clean}{logical; for *MUGA arrays, use only known-good Y chromosome markers}

\item{...}{ignored}
}
\value{
A dataframe of all pre-existing sample metadata with additional columns \code{Ygood} (count of good calls on the
	Y chromosome), \code{Xhet} (count of heterozygous calls on the X chromosome), \code{ix} (mean intensity of X markers) and
	\code{iy} (mean intensity of Y markers).  If no intensity data is availalble, the latter two columns are not included.
}
\description{
Summary stats for the sex chromosomes
}
\details{
If genotypes are not present for *both* sex chromosomes, the results are unpredictable.
}
