% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geno.R
\name{[.genotypes}
\alias{[.genotypes}
\title{Indexing into a \code{genotypes} object}
\usage{
\method{[}{genotypes}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{an object of class \code{genotypes}}

\item{i}{row index (logical, character, or numeric), recycled if needed}

\item{j}{column index (logical, character, or numeric), recycled if needed}

\item{...}{ignored}

\item{drop}{allow reduction of the dimensions of \code{x} if \code{i,j} are of length 1?}
}
\value{
a \code{genotypes} object, preserving any attributes
}
\description{
Allows indexing into a genotype matrix, trimming marker and sample metadata, filters,
and intensity matrices accordingly.
}
\examples{
\dontrun{
geno[ 1:10, ]
geno[ ,c("sample1","sample2") ]
}
}

