% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geno.R
\name{subset.genotypes}
\alias{subset.genotypes}
\title{Subset a \code{genotypes} object by markers or samples}
\usage{
\method{subset}{genotypes}(x, expr, by = c("markers", "samples"), ...)
}
\arguments{
\item{x}{a \code{genotypes} object}

\item{expr}{logical expression indicating elements or rows to keep: missing values are taken as false
(just like \code{subset.data.frame()})}

\item{by}{should \code{expr} be evaluated in the context of markers (eg. by rows) or samples (eg. by columns)?}

\item{...}{ignored}
}
\value{
a \code{genotypes} object similar to \code{x}, but with only the selected rows (\code{by == "markers"})
	or columns (\code{by == "samples"}).  Attributes of \code{x} including intensity matrices, filters,
	marker map and sample metadata are all adjusted accordingly.
}
\description{
Subset a \code{genotypes} object by markers or samples
}
\section{Warning}{

Since this function uses non-standard evaluation, be careful about expressions which include variables defined
in both the global environment and in the scope of the call (eg. variables named same as columns of the marker
map).  In that case results may not be what you expect.
}

