% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{summarize.filters}
\alias{summarize.filters}
\title{Show tally of filters failed by sites, samples.}
\usage{
summarize.filters(gty, filter.codes = "NHIF", ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{filter.codes}{a string of single-character filter codes to include in result}

\item{...}{ignored}
}
\description{
Show tally of filters failed by sites, samples.
}
