% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qc.R
\name{summarize.intensity}
\alias{summarize.intensity}
\title{Summarize hybridization intensity by sample}
\usage{
summarize.intensity(gty, q = seq(0, 1, 0.1), ..., .progress = "none")
}
\arguments{
\item{gty}{a \code{genotypes} object with intensity data attached}

\item{q}{a vector of quantiles (in [0,1])}

\item{.progress}{show a progress bar; passed through to \code{plyr}, see \code{\link{plyr::ddply}}}
}
\value{
a dataframe containing intensity quantiles for each sample, merged with
	sample metadata (if present)
}
\description{
Summarize hybridization intensity by sample
}
\details{
The \code{q}th quantiles of "sum-intensity" are computed for each sample, across all markers
	in the input dataset.  Missing values are silently ignored.  We define "sum-intensity" as
	sqrt(x_i^2 + y_i^2), rather than sum(x_i + y_i), based on the intuition that distance from
	the origin represents total hybridization intensity after application of Illumina's proprietary
	affine-transformation scheme to the raw fluorescences.
}
\seealso{
\code{\link{summarize.calls}}, \code{\link{intensity.vs.ref}}, \code{\link{run.sample.qc}}
}

