% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geno.R
\name{summary.genotypes}
\alias{summary.genotypes}
\title{Show a concise summary of a \code{genotypes} object}
\usage{
\method{summary}{genotypes}(gty, ...)
}
\arguments{
\item{gty}{an object of class \code{genotypes}}

\item{...}{ignored}
}
\description{
Show a concise summary of a \code{genotypes} object
}
\details{
The summary gives the size of the dataset (markers x samples), number of samples
	of each sex, storage mode for alleles, presence of intensity data and whether or not
	intensities have been normalized, and (for files imported with \code{read.beadstudio}) a timestamp
	and checksum.
}

