% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_axesonly}
\alias{theme_axesonly}
\title{A graphics theme with only an x-axis}
\usage{
theme_axesonly(base_size = 12, base_family = "Helvetica")
}
\arguments{
\item{base_size}{baseline font size in points}

\item{base_family}{baseline font family}
}
\value{
a \code{ggplot2} theme function
}
\description{
A graphics theme with only an x-axis
}
\seealso{
\code{\link[ggplot2]{theme}}
}

