% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_gbrowse}
\alias{theme_gbrowse}
\title{A genome-browser like graphics theme}
\usage{
theme_gbrowse(base_size = 12, base_family = "Helvetica")
}
\arguments{
\item{base_size}{baseline font size in points}

\item{base_family}{baseline font family}
}
\value{
a \code{ggplot2} theme function
}
\description{
A genome-browser like graphics theme
}

