% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_heatmap}
\alias{theme_heatmap}
\title{A graphics theme suitable for heatmaps}
\usage{
theme_heatmap(...)
}
\arguments{
\item{...}{ignored}
}
\value{
a \code{ggplot2} theme function
}
\description{
A graphics theme suitable for heatmaps
}
\seealso{
\code{\link[ggplot2]{theme}}
}
