% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_nothing}
\alias{theme_nothing}
\title{A very minimal graphics theme}
\usage{
theme_nothing(base_size = 12, base_family = "Helvetica")
}
\arguments{
\item{base_size}{baseline font size in points}

\item{base_family}{baseline font family}
}
\value{
a \code{ggplot2} theme function
}
\description{
A very minimal graphics theme
}
\seealso{
\code{\link[ggplot2]{theme}}
}
