% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics.R
\name{thin.genotypes}
\alias{thin}
\alias{thin.genotypes}
\title{Thin markers to a specified density on the genetic map}
\usage{
\method{thin}{genotypes}(gty, spacing = 5, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{spacing}{numeric; target distance between retained markers, in cM}

\item{...}{ignored}
}
\value{
a new \code{genotypes} object with fewer markers
}
\description{
Thin markers to a specified density on the genetic map
}

