% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalize.R
\name{tQN}
\alias{tQN}
\title{Perform tQN normalization of intensity data.}
\usage{
tQN(gty, thresholds = c(1.5, 1.5), clusters = NULL, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{thresholds}{thresholds for scaling of x- and y-intensities; defaults recommended in Staaf et al. (2008)}

\item{clusters}{a pre-computed matrix of cluster means}
}
\value{
A copy of the input object, with raw intensities replaced by the normalized ones.  Two additional attributes
\code{baf} and \code{lrr} store the BAF (B-allele frequency) and LRR (log2 intensity ratio).
}
\description{
Perform tQN normalization of intensity data.
}
\details{
Implements thresholded quantile normalization (tQN) as described in Staaf et al. (2008).  Quantile
	normalization as originally described in Bolstad et al. (2003) matches quantiles across multiple samples
	so that all samples' intensities have the same empirical distribution.  The tQN instead matches the quantiles
	of the x- and y-intensities sample-wise in order to reduce noise in the B-allele frequency (BAF) calculation
	proposed by Peiffer et al. (2006).  The quantile-normalized intensities are then subject to a threshold to limit
	on the ratio between the transformed and raw values.
	NB: the quality of the result of tQN depends strongly on the reference clusters provided in \code{clusters},
	so beware.
}
\references{
Adapted from code provided by Johan Staaf.
Staaf J et al. (2008) BMC Bioinformatics. doi:10.1186/1471-2105-9-409.
Bolstad BM et al. (2003) A comparison of normalization methods for high density oligonucleotide
	array data based on bias and variance. Bioinformatics 19(2): 185-193.
Peiffer DA et al. (2006) Genome Res 16(9): 1136-1148. doi:10.1101/gr.5402306.
}

