% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.R
\name{twoway}
\alias{twoway}
\title{Get two-way (marker x marker)-phenotype association}
\usage{
twoway(gty, markers, verbose = TRUE, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{markers}{indexing vector into \code{gty}: can be logical, character or numeric}

\item{verbose}{logical; if \code{TRUE}, preview the contingency table on the console}

\item{...}{ignored}
}
\value{
a dataframe with marker metadata, genotype calls, and sample metadata including phenotype
}
\description{
Get two-way (marker x marker)-phenotype association
}
\details{
A wrapper around \code{oneway()}, generalizing it to marker pairs.  The contingency table
	built from the result of \code{twoway()} will be (genotype x genotype) instead of (genotype x phenotype),
	so the phenotype must be summarizeable.  This function recodes binary phenotypes to 0=control/1=case
	so that taking the mean gives a penetrance value.  Other ordinal or categorical phenotypes will be
	converted from factor to the underlying integer representation, which may or may not be meaningful.
}
