% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plink.R
\name{weir.fst.plink}
\alias{weir.fst.plink}
\title{Compute Weir & Cockerham's F_st estimator using PLINK}
\usage{
weir.fst.plink(prefix, by = NULL, per.locus = FALSE, chr = NULL,
  flags = "", ...)
}
\arguments{
\item{prefix}{a pointer to a PLINK fileset (of class \code{plink})}

\item{by}{a vector of population labels; if not specified, family ID is used}

\item{per.locus}{logical: if \code{TRUE}, return one value per marker; else return aggregate}

\item{chr}{restrict analysis to this chromosome (if \code{NULL}, all chromosomes)}

\item{flags}{command-line flags passed directly to underlying PLINK call}
}
\value{
a square matrix of pairwise F_st values; diagonal has mean heterozygosity within
	each population as defined by the labels in \code{by}
}
\description{
Compute Weir & Cockerham's F_st estimator using PLINK
}
\details{
The F_st statistic as proposed by Sewall Wright is a measure of allele-frequency
	differentiation between two populations. See the relevant PLINK documentation for details
	of the underlying calculations, and Weir & Cockerham (1984) for derivation of the estimator.
}
\references{
PLINK v1.9: \url{https://www.cog-genomics.org/plink2}

Purcell S et al. (2007) PLINK: a toolset for whole-genome association and population-based
	linkage analysis. Am J Hum Genet 81(3): 559-575. doi:10.1086/519795.

Wright S (1931) Evolution in Mendelian populations. Genetics 16: 97-159.

Weir BS & Cockerham CC (1984) Estimating F-statistics for the analysis of population
	structure. Evolution 38(6): 1358-1370.
}

