% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genetics.R
\name{weir.fst}
\alias{weir.fst}
\title{Calculate Weir & Cockerham's unbiased estimator of Fst}
\usage{
weir.fst(gty, subpop = NULL, per.locus = FALSE, ...)
}
\description{
Calculate Weir & Cockerham's unbiased estimator of Fst
}

