% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write.hgdp}
\alias{write.hgdp}
\title{Export genotypes in Stanford HGDP format}
\usage{
write.hgdp(gty, prefix, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{prefix}{filename prefix for output; result will be two files, \code{{prefix}.geno} and
with genotypes and \code{{prefix}.map} with marker map.}

\item{...}{ignored}
}
\description{
Export genotypes in Stanford HGDP format
}
\details{
Write genotypes to disk in the Stanford HGDP format, which can be read by (among
	others) the PGDSpider format-conversion suite.
}
\references{
Lischer HEL and Excoffier L (2012) PGDSpider: An automated data conversion tool for connecting
	population genetics and genomics programs. Bioinformatics 28: 298-299.
}

