% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geno.R
\name{xchrom}
\alias{xchrom}
\title{Shortcut for grabbing just chrX}
\usage{
xchrom(gty, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}
}
\value{
a copy of \code{gty} with only chrX
}
\description{
Shortcut for grabbing just chrX
}

