% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geno.R
\name{ychrom}
\alias{ychrom}
\title{Shortcut for grabbing just chrY}
\usage{
ychrom(gty, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}
}
\value{
a copy of \code{gty} with only chrY
}
\description{
Shortcut for grabbing just chrY
}

