% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qc.R
\name{apply.filters}
\alias{apply.filters}
\title{Drop samples and/or markers flagged as low-quality}
\usage{
apply.filters(gty, apply.to = c("both", "samples", "markers"), ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{apply.to}{dimensions along which to apply filters (samples, sites or both)}
}
\value{
a copy of the input with flagged markers and/or samples dropped
}
\description{
Drop samples and/or markers flagged as low-quality
}
\details{
In the output object, all sample and marker filters are set to \code{FALSE}.
}
\seealso{
\code{\link{run.qc.checks}}, \code{\link{apply.filters}}
}

