% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genetics.R
\name{dist.genotypes}
\alias{dist.genotypes}
\title{Calculate a simple genetic distance: proportion of alleles shared IBS}
\usage{
\method{dist}{genotypes}(gty, ...)
}
\description{
Calculate a simple genetic distance: proportion of alleles shared IBS
}

