% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ex}
\alias{ex}
\title{Genotype calls and intensities from the GigaMUGA array for mouse}
\format{An object of class \code{genotypes} (inherits from \code{matrix}) with 14319 rows and 116 columns.}
\source{
Morgan AP et al. (2015) The Mouse Universal Genotying Array: from substrains
	to subspecies. G3, in press.
}
\usage{
ex
}
\description{
A \code{genotypes} object containing genotype calls and hybridization intensities for
116 control samples (67 male, 49 female) used to calibrate the GigaMUGA array.  To
keep the object at reasonable size, only calls from chr17, chr18, chr19 and chrY
are included (14319 markers).
}
\details{
The samples span 8 inbred mouse strains (n = 58) denoted with 2-letter codes:
\itemize{
\item AA -- A/J
\item BB -- C57BL/6J
\item CC -- 129S1/SvImJ
\item DD -- NOD/ShiLtJ
\item EE -- NZO/HlLtJ
\item FF -- CAST/EiJ
\item GG -- PWK/PhJ
\item HH -- WSB/EiJ
}
The remaining 58 samples are F1s between these 8 strains, also denoted with 2-letter codes
corresponding to their two parental strains.
}
\keyword{datasets}
