% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ex}
\alias{ex}
\alias{inbreds}
\title{Genotype calls and intensities from the GigaMUGA array for mouse}
\format{\preformatted{ genotypes [1:6811, 1:116] 0 0 2 2 2 NA 2 NA NA 2 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:6811] "UNCJPD004122" "JAX00282036" "ICR4265" "JAX00282035" ...
  ..$ : chr [1:116] "AA_0374_F_10004312002_R01C01" "AA_0374_F_10004312002_R11C01" "AA_0417_M_10004312002_R02C01" "AA_0417_M_10004312002_R03C01" ...
 - attr(*, "map")='data.frame':	6811 obs. of  9 variables:
  ..$ chr   : Factor w/ 22 levels "chr1","chr2",..: 10 10 10 10 10 10 10 10 10 10 ...
  ..$ marker: chr [1:6811] "UNCJPD004122" "JAX00282036" "ICR4265" "JAX00282035" ...
  ..$ cM    : num [1:6811] 0.00129 0.00218 0.00219 0.00219 0.00221 ...
  ..$ pos   : int [1:6811] 3102907 3174515 3175147 3175397 3177296 3209050 3216288 3237432 3256308 3262298 ...
  ..$ A1    : chr [1:6811] "T" "T" "T" "A" ...
  ..$ A2    : chr [1:6811] "G" "C" "G" "G" ...
  ..$ type  : Factor w/ 9 levels "CNV_SD","ICR_novel",..: 9 3 2 3 2 4 4 9 9 2 ...
  ..$ is.MM : logi [1:6811] FALSE FALSE FALSE FALSE FALSE FALSE ...
  ..$ unique: logi [1:6811] TRUE TRUE TRUE TRUE TRUE TRUE ...
 - attr(*, "ped")='data.frame':	116 obs. of  6 variables:
  ..$ fid  : chr [1:116] "AA" "AA" "AA" "AA" ...
  ..$ iid  : Factor w/ 116 levels "AA_0374_F_10004312002_R01C01",..: 1 2 3 4 5 6 7 8 9 10 ...
  ..$ mom  : num [1:116] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ dad  : num [1:116] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ sex  : num [1:116] 2 2 1 1 1 2 2 1 2 1 ...
  ..$ pheno: num [1:116] -9 -9 -9 -9 -9 -9 -9 -9 -9 -9 ...
 - attr(*, "intensity")=List of 2
  ..$ x: num [1:6811, 1:116] 0.756 1.853 0.019 0.05 0.018 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:6811] "UNCJPD004122" "JAX00282036" "ICR4265" "JAX00282035" ...
  .. .. ..$ : chr [1:116] "AA_0374_F_10004312002_R01C01" "AA_0374_F_10004312002_R11C01" "AA_0417_M_10004312002_R02C01" "AA_0417_M_10004312002_R03C01" ...
  ..$ y: num [1:6811, 1:116] 0.218 0.037 0.643 1.213 0.682 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:6811] "UNCJPD004122" "JAX00282036" "ICR4265" "JAX00282035" ...
  .. .. ..$ : chr [1:116] "AA_0374_F_10004312002_R01C01" "AA_0374_F_10004312002_R11C01" "AA_0417_M_10004312002_R02C01" "AA_0417_M_10004312002_R03C01" ...
 - attr(*, "normalized")= logi FALSE
 - attr(*, "filter.sites")= Named logi [1:6811] FALSE FALSE FALSE FALSE FALSE FALSE ...
  ..- attr(*, "names")= chr [1:6811] "UNCJPD004122" "JAX00282036" "ICR4265" "JAX00282035" ...
 - attr(*, "filter.samples")= Named logi [1:116] FALSE FALSE FALSE FALSE FALSE FALSE ...
  ..- attr(*, "names")= chr [1:116] "AA_0374_F_10004312002_R01C01" "AA_0374_F_10004312002_R11C01" "AA_0417_M_10004312002_R02C01" "AA_0417_M_10004312002_R03C01" ...
 - attr(*, "class")= chr [1:2] "genotypes" "matrix"
 - attr(*, "alleles")= chr "01"
}}
\source{
Morgan AP et al. (2015) The Mouse Universal Genotying Array: from sister
	strain to sister species. In preparation.
}
\usage{
data(ex); data(inbreds)
}
\description{
A \code{genotypes} object containing genotype calls and hybridization intensities for
116 control samples (67 male, 49 female) used to calibrate the GigaMUGA array.  To
keep the object at reasonable size, only calls from chr10 are included (6811 markers).
}
\details{
The samples span 8 inbred mouse strains (n = 58) denoted with 2-letter codes:
\itemize{
\item AA -- A/J
\item BB -- C57BL/6J
\item CC -- 129S1/SvImJ
\item DD -- NOD/ShiLtJ
\item EE -- NZO/HlLtJ
\item FF -- CAST/EiJ
\item GG -- PWK/PhJ
\item HH -- WSB/EiJ
}
The remaining 58 samples are F1s between these 8 strains, also denoted with 2-letter codes
corresponding to their two parental strains.

The dataset \code{inbreds} contains just the 58 inbred samples.
}
\keyword{datasets}

