% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ex}
\alias{ex}
\alias{inbreds}
\title{Genotype calls and intensities from the GigaMUGA array for mouse}
\format{\preformatted{ chr [1:14319, 1:116] "T" "T" "A" "G" "G" "T" "T" "A" "G" "G" "T" "A" "G" "G" "T" "G" "G" ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:14319] "UNCHS043509" "UNCHS043511" "UNCJPD006513" "UNCHS043510" ...
  ..$ : chr [1:116] "AA_0374_F_10004312002_R01C01" "AA_0374_F_10004312002_R11C01" "AA_0417_M_10004312002_R02C01" "AA_0417_M_10004312002_R03C01" ...
 - attr(*, "map")='data.frame':	14319 obs. of  10 variables:
  ..$ chr         : Factor w/ 22 levels "chr1","chr2",..: 17 17 17 17 17 17 17 17 17 17 ...
  ..$ marker      : chr [1:14319] "UNCHS043509" "UNCHS043511" "UNCJPD006513" "UNCHS043510" ...
  ..$ cM          : num [1:14319] 0.00663 0.00663 0.0071 0.00737 0.00737 ...
  ..$ pos         : int [1:14319] 3075928 3075928 3081326 3084336 3084336 3087345 3159640 3161972 3164218 3166961 ...
  ..$ A1          : chr [1:14319] "T" "T" "A" "T" ...
  ..$ A2          : chr [1:14319] "C" "C" "G" "G" ...
  ..$ type        : Factor w/ 9 levels "CNV_SD","ICR_novel",..: 4 4 9 4 4 3 4 4 4 4 ...
  ..$ is.MM       : logi [1:14319] FALSE FALSE FALSE FALSE FALSE FALSE ...
  ..$ unique      : logi [1:14319] TRUE TRUE TRUE TRUE TRUE TRUE ...
  ..$ is.biallelic: logi [1:14319] TRUE TRUE TRUE TRUE TRUE TRUE ...
 - attr(*, "ped")='data.frame':	116 obs. of  6 variables:
  ..$ fid  : chr [1:116] "inbred" "inbred" "inbred" "inbred" ...
  ..$ iid  : Factor w/ 116 levels "AA_0374_F_10004312002_R01C01",..: 1 2 3 4 5 6 7 8 9 10 ...
  ..$ mom  : num [1:116] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ dad  : num [1:116] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ sex  : num [1:116] 2 2 1 1 1 2 2 1 2 1 ...
  ..$ pheno: int [1:116] 2 2 2 2 2 2 2 2 2 1 ...
 - attr(*, "alleles")= chr "native"
 - attr(*, "filter.sites")= Named chr [1:14319] "" "" "" "" ...
  ..- attr(*, "names")= chr [1:14319] "UNCHS043509" "UNCHS043511" "UNCJPD006513" "UNCHS043510" ...
 - attr(*, "filter.samples")= Named chr [1:116] "" "" "" "" ...
  ..- attr(*, "names")= chr [1:116] "AA_0374_F_10004312002_R01C01" "AA_0374_F_10004312002_R11C01" "AA_0417_M_10004312002_R02C01" "AA_0417_M_10004312002_R03C01" ...
 - attr(*, "class")= chr [1:2] "genotypes" "matrix"
 - attr(*, "intensity")=List of 2
  ..$ x: num [1:14319, 1:116] 1.247 1.095 1.021 0.253 0.544 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:14319] "UNCHS043509" "UNCHS043511" "UNCJPD006513" "UNCHS043510" ...
  .. .. ..$ : chr [1:116] "AA_0374_F_10004312002_R01C01" "AA_0374_F_10004312002_R11C01" "AA_0417_M_10004312002_R02C01" "AA_0417_M_10004312002_R03C01" ...
  ..$ y: num [1:14319, 1:116] 0.067 0.053 0.208 1.067 1.512 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:14319] "UNCHS043509" "UNCHS043511" "UNCJPD006513" "UNCHS043510" ...
  .. .. ..$ : chr [1:116] "AA_0374_F_10004312002_R01C01" "AA_0374_F_10004312002_R11C01" "AA_0417_M_10004312002_R02C01" "AA_0417_M_10004312002_R03C01" ...
 - attr(*, "normalized")= logi FALSE
}}
\source{
Morgan AP et al. (2015) The Mouse Universal Genotying Array: from sister
	strain to sister species. In preparation.
}
\usage{
data(ex); data(inbreds)
}
\description{
A \code{genotypes} object containing genotype calls and hybridization intensities for
116 control samples (67 male, 49 female) used to calibrate the GigaMUGA array.  To
keep the object at reasonable size, only calls from chr10 are included (6811 markers).
}
\details{
The samples span 8 inbred mouse strains (n = 58) denoted with 2-letter codes:
\itemize{
\item AA -- A/J
\item BB -- C57BL/6J
\item CC -- 129S1/SvImJ
\item DD -- NOD/ShiLtJ
\item EE -- NZO/HlLtJ
\item FF -- CAST/EiJ
\item GG -- PWK/PhJ
\item HH -- WSB/EiJ
}
The remaining 58 samples are F1s between these 8 strains, also denoted with 2-letter codes
corresponding to their two parental strains.

The dataset \code{inbreds} contains just the 58 inbred samples.
}
\keyword{datasets}

