% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics.R
\name{freq}
\alias{freq}
\title{Calculate allele frequencies by marker or sample}
\usage{
freq(gty, by = c("markers", "samples"), na.rm = TRUE, counts = FALSE, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object with numeric allele encoding}

\item{by}{margin over which to calculate frequencies (usually \code{"markers"})}

\item{na.rm}{skip over missing genotypes in frequency calculations}

\item{counts}{logical; if \code{TRUE}, report allele counts instead of relative frequencies}

\item{...}{ignored}
}
\value{
a named vector of allele frequencies
}
\description{
Calculate allele frequencies by marker or sample
}
