% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geno.R
\name{head.genotypes}
\alias{head.genotypes}
\title{Show the first few samples and markers for a \code{genotypes}}
\usage{
head.genotypes(x, n = 10, nsamples = min(10, ncol(x)), ...)
}
\arguments{
\item{x}{an object of class \code{genotypes}}

\item{n}{number of markers to show}

\item{nsamples}{number of samples to show}

\item{...}{ignored}
}
\description{
Show the first few samples and markers for a \code{genotypes}
}
\details{
Sample names are abbreviated to keep output tidy.
}

