% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{heatmap.genotypes}
\alias{heatmap}
\alias{heatmap.genotypes}
\title{Plot a heatmap representing genetic distances between samples}
\usage{
\method{heatmap}{genotypes}(gty, d = NULL, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{d}{a distance matrix (of class \code{dist}), to avoid re-computing}

\item{...}{ignored}
}
\value{
a (hierarchically-clustered) matrix representation of pairwise genetic distances between
	samples, normalized to [0,1].
}
\description{
Plot a heatmap representing genetic distances between samples
}
\details{
Computation of pairwise distances in a large set of samples will be somewhat expensive,
	so there is advantage to pre-computing this matrix (using \code{argyle::dist()}) if it may
	be useful in other analyses.
}

