% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genetics.R
\name{heterozygosity}
\alias{heterozygosity}
\title{Calculate heterozygosity by marker}
\usage{
heterozygosity(gty, hwe = FALSE, na.rm = TRUE, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object}

\item{hwe}{if \code{TRUE}, compute the MLE \code{hat{h}} assuming Hardy-Weinberg equilibrium}

\item{na.rm}{logical; ignore missing genotypes?}

\item{...}{ignored}
}
\value{
a named vector with the heterozygosity of each marker
}
\description{
Calculate heterozygosity by marker
}

