% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geno.R
\name{merge.genotypes}
\alias{merge.genotypes}
\title{Merge two \code{genotypes} objects which share markers}
\usage{
\method{merge}{genotypes}(a, b, join = c("inner", "left"), ...)
}
\arguments{
\item{a}{a \code{genotypes} object}

\item{b}{another \code{genotypes} object}

\item{join}{what sort of join to perform: at present, only \code{"inner"} (intersection of \code{a}'s and \code{b}'s
    marker sets) is supported}
}
\value{
a \code{genotypes} object containing all samples in \code{a} and \code{b}, at all markers shared
	by \code{a} and \code{b}
}
\description{
Merge two \code{genotypes} objects which share markers
}
\details{
Both input objects should have the same allele coding (character vs. numeric, relative vs. absolute)
	before merging.  Sharing of markers is established by the presence of overlapping rownames in the genotype
	matrices of the input, but alleles are NOT checked.  Marker names must match exactly if they are to appear
	in the output.  Intensity matrices, if present, are merged and carried over to the output. They will be marked
	as normalized only if they are marked as such in both input objects.  Vectors of site and sample filters will
	be present in the output only if they are present in both input objects, and have the expected names.  A
	logical OR is applied to the site filters -- ie. a site will be marked as \code{TRUE} (filtered) if it is
	filtered in either or both input datasets.
}

