% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{read.plink}
\alias{read.plink}
\title{Read a PLINK binary fileset into a \code{genotypes} object}
\usage{
read.plink(prefix, ...)
}
\arguments{
\item{prefix}{path to a PLINK fileset, excluding \code{*.bed} suffix}

\item{...}{ignored}
}
\value{
a \code{genotypes} object, with alleles in the "01" encoding (see \code{\link{recode.genotypes}})
}
\description{
Read a PLINK binary fileset into a \code{genotypes} object
}
\details{
Reads a PLINK binary fileset using pure \code{R}.  The fileset should be generated in 
	the "variant-major" format (the default in all recent versions of PLINK.) Missing genotypes are
	represented as \code{NA}s.
}
\references{
PLINK v1.9: \url{https://www.cog-genomics.org/plink2}

Purcell S et al. (2007) PLINK: a toolset for whole-genome association and population-based 
	linkage analysis. Am J Hum Genet 81(3): 559-575. doi:10.1086/519795.
}
\seealso{
\code{\link{write.plink}}
}

