% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{summarize.calls}
\alias{summarize.calls}
\title{Summarize genotype calls by sample or marker}
\usage{
summarize.calls(gty, by = c("samples", "markers"), counts = TRUE, ...)
}
\arguments{
\item{gty}{a \code{genotypes} object with intensity data attached}

\item{by}{get call rates by sample or by marker}

\item{counts}{logical; if \code{TRUE}, return absolute counts, otherwise relative frequencies}

\item{...}{ignored}
}
\value{
a dataframe with counts of A (reference or major allele), B (alternate or minor allele),
	H (heterozygous) and N (missing/no-call) by either sample or marker.
}
\description{
Summarize genotype calls by sample or marker
}
\details{
Any metadata associated with markers or samples is *not* merged into the final result, in order
	to preserve parallelism between rows of the result and of the parent object.
}
\seealso{
\code{\link{summarize.intensity}}, \code{\link{intensity.vs.ref}}, \code{\link{run.sample.qc}}
}

