% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plink.R
\name{write.plink}
\alias{write.plink}
\title{Write a \code{genotypes} object as a PLINK binary fileset}
\usage{
write.plink(gty, prefix, map = NULL, fam = NULL, ...)
}
\arguments{
\item{gty}{an \code{genotypes} object}

\item{prefix}{path to the PLINK fileset to generate, excluding \code{*.bed} suffix}

\item{map}{a valid marker map; overrides existing map}

\item{fam}{sample metadata to override any existing metadata}

\item{...}{ignored}
}
\value{
\code{TRUE} on completion
}
\description{
Write a \code{genotypes} object as a PLINK binary fileset
}
\details{
Writes a PLINK binary fileset using pure \code{R}.  The fileset is generated in the
	"variant-major" format (the default in all recent versions of PLINK), with appropriate magic
	number.  Complete sample metadata and a complete marker map, including reference alleles (columns "A1"
	and "A2"), are both required.

	This function only supports writing from character (allele enconding \code{"native"}) genotypes,
	in order to avoid ambiguity around the correspondence between alleles and numeric genotypes.  But
	numeric genotypes can be converted back to character via \code{recode.genotypes(,"native")}.
}
\references{
PLINK v1.9: \url{https://www.cog-genomics.org/plink2}

Purcell S et al. (2007) PLINK: a toolset for whole-genome association and population-based
	linkage analysis. Am J Hum Genet 81(3): 559-575. doi:10.1086/519795.
}
\seealso{
\code{\link{read.plink}}
}

